feat: add grocery list items service with create method

+++ src/main/java/com/om1cael/fryna/api/data/dtos/GroceryListItemDTO.java
@@ -19,6 +19,6 @@ public record GroceryListItemDTO(
     boolean bought,
 
     @NotNull
-    int listID
+    Long listID
 ) {
 }
